pyproject.toml
[tool.poetry]
name = "coatCareAiTest"
version = "0.1.0"
description = "coat care ai test"
authors = ["denis"]
repository = ""

[tool.poetry.dependencies]
python = "^3.12"
pydantic = { version = "^2.7.4", extras = ["email"] }
pydantic-settings = "^2.12.0"
python-dotenv = "^1.2.1"
google-api-python-client = "^2.189.0"
google-auth-httplib2 = "^0.2.1"
google-auth-oauthlib = "^1.2.4"
langchain = "^1.2.8"
langchain-google-genai = "^4.2.0"
langchain-community = "^0.4.1"
langchain-chroma = "^1.1.0"
ipython = "^9.10.0"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^6.0.0"
pytest-cov = "^4.1.0"
bandit = "^1.7.5"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.black]
line_length = 120

[tool.isort]
profile = "black"
line_length = 120
